feat: naive time quanta based scheduler
Abstracted the current program as a `main` go routine in preparation for
`GoStatement` implementation
